% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{in_pkgdown}
\alias{in_pkgdown}
\title{Determine if the package is being used by pkgdown}
\usage{
in_pkgdown()
}
\value{
a logical value (\code{TRUE} or \code{FALSE}).
}
\description{
This function should not be called by the user.
It allows to run some examples conditionally to being used by pkgdown.
Code copied from \code{\link[pkgdown:in_pkgdown]{pkgdown::in_pkgdown()}}.
}
