% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI}
\alias{CI}
\alias{calc_H}
\alias{calc_e}
\alias{calc_Gamma}
\alias{calc_GammaINF}
\alias{calc_L_And}
\alias{calc_D}
\alias{calc_SE}
\title{Internal functions for the computation of confidence intervals}
\usage{
calc_H(smoothed_obj)

calc_e(smoothed_obj, H)

calc_Gamma(e, l)

calc_GammaINF(e, L)

calc_L_And(e, AR.method = c("yule-walker", "burg", "ols", "mle", "yw"))

calc_D(smoothed_obj)

calc_SE(
  smoothed_obj,
  h,
  AR.method = c("yule-walker", "burg", "ols", "mle", "yw")
)
}
\arguments{
\item{smoothed_obj}{an object created with \code{\link{calc_rho}}.}

\item{H}{an object created with \code{\link{calc_H}}.}

\item{e}{an object created with \code{\link{calc_e}}.}

\item{l}{a scalar indicating a number of time points.}

\item{L}{a scalar indicating a bandwidth parameter.}

\item{AR.method}{character string specifying the method to fit the autoregressive model used to compute \verb{$\\hat\{\\gamma\}_1$} in \verb{$L_\{And\}$} (see \code{\link[stats:ar]{stats::ar}} for details).}

\item{h}{a scalar indicating the bandwidth used by the smoothing function.}
}
\description{
These functions compute the different terms required to compute the confidence
interval around the time-varying correlation coefficient.
}
\section{Functions}{
\itemize{
\item \code{calc_H()}: Internal function computing the \verb{$\\hat\{H_t\}$} array.

\verb{$\\hat\{H_t\}$} is a component needed to compute confidence intervals;
\verb{$H_t$} is defined in eq. 6 from Choi & Shin, 2021.
The function returns a 5 x 5 x \code{t} array.

\item \code{calc_e()}: Internal function computing \verb{$\\hat\{e\}_t$}.

\verb{$\\hat\{e\}_t$} is a component needed to compute confidence intervals;
it is defined in eq. 9 from Choi & Shin, 2021.
The function returns a \code{t} x 5 matrix storing the residuals.

\item \code{calc_Gamma()}: Internal function computing \verb{$\\hat\{\\Gamma\}_l$}.

\verb{$\\hat\{\\Gamma\}_l$} is a component needed to compute confidence intervals;
it is defined in eq. 9 from Choi & Shin, 2021.
The function returns a 5 x 5 matrix.

\item \code{calc_GammaINF()}: Internal function computing \verb{$\\hat\{\\Gamma\}^\\Inf$}.

\verb{$\\hat\{\\Gamma\}^\\Inf$} is a component needed to compute confidence intervals (the long run variance estimator);
it is defined in eq. 9 from Choi & Shin, 2021.
The function returns a 5 x 5 matrix.

\item \code{calc_L_And()}: Internal function computing \verb{$L_\{And\}$}.

\verb{$L_\{And\}$} is a component needed to compute confidence intervals;
it is defined in Choi & Shin, 2021, p 342.
It also corresponds to \verb{$S_T^*$}, eq 5.3 in Andrews 1991.
The function returns a scalar which should be used as an input for \code{L} in \code{\link{calc_GammaINF}}.

\item \code{calc_D()}: Internal function computing \verb{$D_t$}.

\verb{$D_t$} is a component needed to compute confidence intervals;
it is defined in Choi & Shin, 2021, p 338.
The function returns a \code{t} x 5 matrix storing the residuals.

\item \code{calc_SE()}: Internal function computing \verb{$se(\\hat\{rho\}_t(h))$}.

The standard deviation of the time-varying correlation (\verb{$se(\\hat\{rho\}_t(h))$}) is defined in eq. 8 from Choi & Shin, 2021.
It depends on \verb{$D_\{Lt\}$}, \verb{$D_\{Mt\}$} & \verb{$D_\{Ut\}$}, themselves defined in Choi & Shin, 2021, p 337 & 339.
The \verb{$D_\{Xt\}$} terms are all computed within the function since they all rely on the same components.
The function returns a vector of length \code{t}.

}}
\examples{
rho_obj <- with(na.omit(stockprice),
                calc_rho(x = SP500, y = FTSE100, t = DateID, h = 20, kernel = "box"))
head(rho_obj)


## Computing `$\hat{H_t}$`

H <- calc_H(smoothed_obj = rho_obj)
H[, , 1:2] # H array for the first two time points


## Computing `$\hat{e}_t$`

e <- calc_e(smoothed_obj = rho_obj, H = H)
head(e) # e matrix for the first six time points


## Computing `$\hat{\Gamma}_l$`

calc_Gamma(e = e, l = 3)


## Computing `$\hat{\Gamma}^\Inf$`

calc_GammaINF(e = e, L = 2)


## Computing `$L_{And}$`

calc_L_And(e = e)
sapply(c("yule-walker", "burg", "ols", "mle", "yw"),
       function(m) calc_L_And(e = e, AR.method = m)) ## comparing AR.methods


## Computing `$D_t$`

D <- calc_D(smoothed_obj = rho_obj)
head(D) # D matrix for the first six time points


run <- FALSE ## change to TRUE to run the example
if (in_pkgdown() || run) {

## Computing `$se(\hat{rho}_t(h))$`
# nb: takes a few seconds to run

SE <- calc_SE(smoothed_obj = rho_obj, h = 50)
head(SE) # SE vector for the first six time points

}


}
\references{
Choi, JE., Shin, D.W. Nonparametric estimation of time varying correlation coefficient.
J. Korean Stat. Soc. 50, 333–353 (2021). \doi{10.1007/s42952-020-00073-6}

Andrews, D. W. K. Heteroskedasticity and autocorrelation consistent covariance matrix estimation.
Econometrica: Journal of the Econometric Society, 817-858 (1991).
}
\seealso{
\code{\link{tcor}}
}
