% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI}
\alias{CI}
\alias{calc_H}
\alias{calc_e}
\alias{calc_Gamma}
\alias{calc_GammaINF}
\alias{calc_L_And}
\title{Internal functions for the computation of confidence intervals}
\usage{
calc_H(smoothed_obj)

calc_e(smoothed_obj, H)

calc_Gamma(e, l)

calc_GammaINF(e, L)

calc_L_And(e, AR.method = c("yule-walker", "burg", "ols", "mle", "yw"))
}
\arguments{
\item{smoothed_obj}{an object created with \code{\link{calc_rho}}.}

\item{H}{an object created with \code{\link{calc_H}}.}

\item{e}{an object created with \code{\link{calc_e}}.}

\item{l}{a scalar indicating a number of time steps.}

\item{L}{a scalar indicating a bandwidth parameter.}

\item{AR.method}{character string specifying the method to fit the autoregressive model used to compute \verb{$\\hat\{\\gamma\}_1$} in \verb{$L_\{And\}$} (see \code{\link[stats:ar]{stats::ar}} for details).}
}
\description{
These functions compute the different terms required to compute the confidence
interval around the time-varying correlation coefficients.
}
\section{Functions}{
\itemize{
\item \code{calc_H}: Internal function computing the \verb{$\\hat\{H_t\}$} array.

\verb{$\\hat\{H_t\}$} is a component needed to compute confidence intervals;
\verb{$H_t$} is defined in eq. 6 from Choi & Shin, 2021.
The function returns a 5 x 5 x t array.

\item \code{calc_e}: Internal function computing \verb{$\\hat\{e\}_t$}.

\verb{$\\hat\{e\}_t$} is a component needed to compute confidence intervals;
it is defined in eq. 9 from Choi & Shin, 2021.
The function returns a 5 x t matrix storing the residuals.

\item \code{calc_Gamma}: Internal function computing \verb{$\\hat\{\\Gamma\}_l$}.

\verb{$\\hat\{\\Gamma\}_l$} is a component needed to compute confidence intervals;
it is defined in eq. 9 from Choi & Shin, 2021.
The function returns a 5 x 5 matrix.

\item \code{calc_GammaINF}: Internal function computing \verb{$\\hat\{\\Gamma\}^\\Inf$}.

\verb{$\\hat\{\\Gamma\}^\\Inf$} is a component needed to compute confidence intervals (the long run variance estimator);
it is defined in eq. 9 from Choi & Shin, 2021.
The function returns a 5 x 5 matrix.

\item \code{calc_L_And}: Internal function computing \verb{$L_\{And\}$}.

\verb{$L_\{And\}$} is a component needed to compute confidence intervals;
it is defined in Choi & Shin, 2021, p 342.
It also corresponds to \verb{$S_T^*$}, eq 5.3 in Andrews 1991.
The function returns a scalar which should be used as an input for \code{L} in \code{\link{calc_GammaINF}}.
}}

\examples{

rho_obj <- with(na.omit(stockprice),
                calc_rho(x = SP500, y = FTSE100, t = DateID, h = 20, kernel = "box"))
head(rho_obj)

## Computing `$\hat{H_t}$`

H <- calc_H(smoothed_obj = rho_obj)
H

## Computing `$\hat{e}_t$`

e <- calc_e(smoothed_obj = rho_obj, H = H)
e

## Computing `$\hat{\Gamma}_l$`

calc_Gamma(e = e, l = 3)

## Computing `$\hat{\Gamma}^\Inf$`

calc_GammaINF(e = e, L = 2)

## Computing `$L_{And}$`

calc_L_And(e = e)

}
\references{
Choi, JE., Shin, D.W. Nonparametric estimation of time varying correlation coefficient.
J. Korean Stat. Soc. 50, 333–353 (2021). https://doi.org/10.1007/s42952-020-00073-6

Andrews, D. W. K. Heteroskedasticity and autocorrelation consistent covariance matrix estimation.
Econometrica: Journal of the Econometric Society, 817-858 (1991).
}
\seealso{
\code{\link{tcor}}
}
